[test] update beaconchain5.txt file
improve deploy.sh, don't download key if exists

+++ test/deploy.sh
@@ -175,7 +175,9 @@ while IFS='' read -r line || [[ -n "$line" ]]; do
         then
           echo ""${blspub}.key" already in local."
        else
-          aws s3 cp "s3://harmony-secret-keys/bls-test/${blspub}.key" .hmy
+          if [ ! -e .hmy/${blspub}.key ]; then
+             aws s3 cp "s3://harmony-secret-keys/bls-test/${blspub}.key" .hmy
+          fi
       fi
 
       args=("${base_args[@]}" -ip "${ip}" -port "${port}" -key "/tmp/${ip}-${port}.key" -db_dir "db-${ip}-${port}" -accounts "${account}" -blspass file:blspass.txt -blskey_file ".hmy/${blspub}.key")
